Add support for LMU
Weather generation does not work for now because there seems to be
a bug with LMU: it does not load those .wet files for race sessions
while still writing to them. Also, even if you configure weather
manually through the UI - it will work for the current race but for
the next one (even without closing the game) it will pick a new random
weather.